pmath (Linux x64):   - pmath_build_value(): using va_copy() where needed   - removed fmod() from _pmath_event_timedwait() becaused that segfaulted for some reason   - removed PMATH_INT_32BIT/ PMATH_INT_64BIT macros   - fixed pmath_object_atomic_write()
richmath:
  - some changes for older gtk/gdk versions

+++ richmath/include/util/sharedptr.h
@@ -1,10 +1,12 @@
 #ifndef __UTIL__SHAREDPTR_H__
 #define __UTIL__SHAREDPTR_H__
 
+#include <stdint.h>
+
 
 namespace richmath {
   void assert_failed();
-  
+
   /* T must provide public methods void ref() and void unref().
      The constructor of T should set its reference counter to 1. */
   template<class T> class SharedPtr {
@@ -13,40 +15,40 @@ namespace richmath {
         : _ptr(ptr)
       {
       }
-      
+
       SharedPtr(const SharedPtr<T> &src)
         : _ptr(src._ptr)
       {
         if(_ptr)
           _ptr->ref();
       }
-      
+
       template<class O> SharedPtr(const SharedPtr<O> &src)
         : _ptr(dynamic_cast<T*>(src.ptr()))
       {
         if(_ptr)
           _ptr->ref();
       }
-      
+
       ~SharedPtr() {
         if(_ptr)
           _ptr->unref();
       }
-      
+
       SharedPtr &operator=(T *ptr) {
         if(_ptr)
           _ptr->unref();
         _ptr = ptr;
         return *this;
       }
-      
+
       SharedPtr &operator=(const SharedPtr &src) {
         /* In bla = bla->next, where bla._ptr.refcount = 1, _ptr would be
            deleted and thus src === &bla->next === &bla._ptr->next is lost, so
            we need to preserve the value in a local variable.
          */
         T *p = src._ptr;
-        
+
         if(p)
           p->ref();
         if(_ptr)
@@ -54,45 +56,45 @@ namespace richmath {
         _ptr = p;
         return *this;
       }
-      
+
       //const T *ptr() const { return _ptr; }
-      
+
       T *ptr() const { return _ptr; }
-      
+
       T *release() {
         T *result = _ptr;
         _ptr = 0;
         return result;
       }
-      
+
       const T *operator->() const {
         return _ptr;
       }
-      
+
       T *operator->() {
         return _ptr;
       }
-      
+
       template<class O> bool operator==(const SharedPtr<O> &src) const {
         return _ptr == src._ptr;
       }
-      
+
       template<class O> bool operator!=(const SharedPtr<O> &src) const {
         return _ptr != src._ptr;
       }
-      
+
       bool is_valid() const {
         return _ptr != 0;
       }
-      
+
       operator bool() const {
         return is_valid();
       }
-      
+
       unsigned int hash() const {
-        return (unsigned int)(const void*)_ptr;
+        return (unsigned int)(uintptr_t)(const void*)_ptr;
       }
-      
+
     protected:
       T *_ptr;
   };
